Added in level1 duplicate as level3 to playtest cached score.  Cached score is now implemented.
Everytime loads a new level that isn't final it sets a variable total score to the score currently displayed (BP_GameMode). In HUD setting an int variable of the current score to set the total score to. Also everytime it updates the score in the HUD it adds the current levels score to the total score. Adjusted comments and wires organization.